Add support for wavedrom images
This adds support for wavedrom images, instead of inline
javascript. By this it supports other builders as html, in particular
latexpdf (while others may be added too).

Also for html builders, images can be used by setting the
configuration option

    wavedrom_html_jsinline = False

This setting defaults to True to not change the default behavior.

As a caveat, this only works really nice with Python 3 (for other
output formats then svg), because it uses cairosvg to generate pdf and
png. But at least for html that is fine, and has the advantage that it
doesn't depend on any command line tools.